Flatten file structure where there is only one execution model
