--- src/capture/mod.rs
@@ -1,3 +1,0 @@
-//! Input capture traits
-
-pub mod nb;
