--- src/delay/mod.rs
@@ -1,3 +1,0 @@
-//! Delay traits
-
-pub mod blocking;
