--- src/i2c/mod.rs
@@ -1,20 +1,0 @@
-//! I2C traits
-
-pub mod blocking;
-
-use crate::private;
-
-/// Address mode (7-bit / 10-bit)
-///
-/// Note: This trait is sealed and should not be implemented outside of this crate.
-pub trait AddressMode: private::Sealed {}
-
-/// 7-bit address mode type
-pub type SevenBitAddress = u8;
-
-/// 10-bit address mode type
-pub type TenBitAddress = u16;
-
-impl AddressMode for SevenBitAddress {}
-
-impl AddressMode for TenBitAddress {}
