--- src/pwm/mod.rs
@@ -1,3 +1,0 @@
-//! Pulse width modulation traits
-
-pub mod blocking;
