--- src/timer/mod.rs
@@ -1,6 +1,0 @@
-//! Timer traits
-
-pub mod nb;
-
-/// Marker trait that indicates that a timer is periodic
-pub trait Periodic {}
