--- src/watchdog/mod.rs
@@ -1,3 +1,0 @@
-//! Watchdog traits
-
-pub mod blocking;
